# internal/models/responses.go
package models

type TokenResponse struct {
    Token string `json:"token" example:"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."`
}

type ErrorResponse struct {
    Error string `json:"error" example:"error message"`
}	